requirements.txt
APScheduler==3.9.1
Flask==2.2.2
pytest==7.1.2
PyYAML==6.0
requests==2.28.1
